requirements.txt
pandas>=2.0.0
pyyaml>=6.0
sqlalchemy>=2.0.0
pymongo>=4.0.0
requests>=2.28.0
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pyarrow>=12.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
office365-rest-python-client>=2.4.0
psycopg2-binary>=2.9.0
mysql-connector-python>=8.0.0